refactor(JSAP-4): refactor frontend tests
