+++ frontend/package.json
@@ -34,7 +34,7 @@
     "eslint-plugin-react": "^7.16.0",
     "eslint-plugin-react-hooks": "^1.7.0",
     "jest-expo": "^35.0.0",
-    "react-test-renderer": "^16.8.6"
+    "react-test-renderer": "^16.8.3"
   },
   "jest": {
     "preset": "jest-expo",
